sua meal record

+++ src/main/java/pidvn/modules/hr/meal/services/HrMealSvc.java
@@ -3,19 +3,13 @@ package pidvn.modules.hr.meal.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import pidvn.entities.one.EMealData;
-import pidvn.entities.one.HrAttendanceDetail;
-import pidvn.entities.one.HrMealRecord;
-import pidvn.entities.one.HrOvertimeData;
+import pidvn.entities.one.*;
 import pidvn.mappers.one.hr.e_meal.HrEMealMapper;
 import pidvn.mappers.three.hr.meal.HrMealMapper;
 import pidvn.modules.hr.meal.models.MealCouponVo;
 import pidvn.modules.hr.meal.models.MealRecordVo;
 import pidvn.modules.hr.meal.models.SearchVo;
-import pidvn.repositories.one.EMealDataRepo;
-import pidvn.repositories.one.HrAttendanceDetailRepo;
-import pidvn.repositories.one.HrMealRecordRepo;
-import pidvn.repositories.one.HrOvertimeDataRepo;
+import pidvn.repositories.one.*;
 
 import javax.mail.*;
 import javax.mail.internet.InternetAddress;
@@ -43,6 +37,9 @@ public class HrMealSvc implements IHrMealSvc {
     @Autowired
     private HrMealRecordRepo hrMealRecordRepo;
 
+    @Autowired
+    private HrLeaveDayRepo hrLeaveDayRepo;
+
 
     @Override
     public Map getMealRecords(SearchVo searchVo) {
@@ -65,47 +62,65 @@ public class HrMealSvc implements IHrMealSvc {
 
     @Override
     @Transactional(transactionManager = "transactionManagerOne")
-    public Map timesheetConfirm() {
+    public Map timesheetConfirm(String table) {
 
-        /**
-         * Xóa data AttendanceDetails tháng trước đó ở FDSC db
-         */
-        this.hrAttendanceDetailRepo.deleteAttendanceDetailsPreviousMonth();
+        Map result = new HashMap();
 
-        /**
-         * Xóa data Overtime tháng trước đó ở FDSC db
-         */
-        this.hrOvertimeDataRepo.deleteOvertimeDataPreviousMonth();
+        if (table.equals("attendance")) {
+            // Lấy data từ PVG
+            List<HrAttendanceDetail> attendanceDetails = this.hrMealMapper.getAttendanceDetails();
 
-        /**
-         * Xóa data MealRecord tháng trước đó ở FDCS db
-         */
-        this.hrMealRecordRepo.deleteMealRecordsPreviousMonth();
+            // Xóa data
+            this.hrAttendanceDetailRepo.deleteAttendanceDetailsPreviousMonth();
 
-        /**
-         * Lấy dữ liệu AttendanceDetails, Overtime, MealRecord từ PVG database
-         */
-        List<HrAttendanceDetail> attendanceDetails = this.hrMealMapper.getAttendanceDetails();
-        List<HrOvertimeData> overtimeData = this.hrMealMapper.getOvertimeData();
-        List<HrMealRecord> mealRecords = this.hrMealMapper.getHrMealRecord();
+            // Lưu data vào MySQL
+            List<HrAttendanceDetail> attendanceDetailResult = this.hrAttendanceDetailRepo.saveAll(attendanceDetails);
 
-        /**
-         * Lưu dữ liệu AttendanceDetails, Overtime, MealRecord vào FDCS database
-         */
-        List<HrAttendanceDetail> attendanceDetailResult = this.hrAttendanceDetailRepo.saveAll(attendanceDetails);
-        List<HrOvertimeData> overtimeResult = this.hrOvertimeDataRepo.saveAll(overtimeData);
-        List<HrMealRecord> mealRecordResult = this.hrMealRecordRepo.saveAll(mealRecords);
+            result.put("table", "attendance");
+            result.put("record", attendanceDetailResult.size());
+        }
 
-        /**
-         * Lưu dữ liệu MealRecords
-         */
+        if (table.equals("overtime")) {
+            // Lấy data từ PVG
+            List<HrOvertimeData> overtimeData = this.hrMealMapper.getOvertimeData();
 
-        Map map = new HashMap();
-        map.put("attendanceDetailResult", attendanceDetailResult.size());
-        map.put("overtimeResult", overtimeResult.size());
-        map.put("mealRecordResult", mealRecordResult.size());
+             // Xóa data
+            this.hrOvertimeDataRepo.deleteOvertimeDataPreviousMonth();
 
+            // Lưu data vào MySQL
+            List<HrOvertimeData> overtimeResult = this.hrOvertimeDataRepo.saveAll(overtimeData);
+            result.put("table", "overtime");
+            result.put("record", overtimeResult.size());
+        }
 
-        return map;
+        if (table.equals("meal_record")) {
+            // Lấy data từ PVG
+            List<HrMealRecord> mealRecords = this.hrMealMapper.getHrMealRecord();
+
+            // Xóa data
+            this.hrMealRecordRepo.deleteMealRecordsPreviousMonth();
+
+            // Lưu data vào MySQL
+            List<HrMealRecord> mealRecordResult = this.hrMealRecordRepo.saveAll(mealRecords);
+
+            result.put("table", "meal_record");
+            result.put("record", mealRecordResult.size());
+        }
+
+        if(table.equals("leave_day")) {
+            // Lấy data từ PVG
+            List<HrLeaveDay> leaveDays = this.hrMealMapper.getHrLeaveDay();
+
+            // Xóa data
+            this.hrLeaveDayRepo.deleteLeaveDayPreviousMonth();
+
+            // Lưu data vào MySQL
+            List<HrLeaveDay> leaveDaysResult = this.hrLeaveDayRepo.saveAll(leaveDays);
+
+            result.put("table", "leave_day");
+            result.put("record", leaveDaysResult.size());
+        }
+
+        return result;
     }
 }
